# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface DashboardSummary {
  workoutsThisWeek: number;
  streakDays: number;
  caloriesToday: number;
  calorieGoal: number;
  proteinToday: number;
  carbsToday: number;
  fatToday: number;
  proteinGoal: number;
  carbGoal: number;
  fatGoal: number;
  totalPoints: number;
  name: string;
  /** @nullable */
  imageUrl: string | null;
}

export interface WeekScheduleEntry {
  id: string;
  date: string;
  title: string;
  completed: boolean;
  /** @nullable */
  source?: string | null;
}

export interface ActiveChallenge {
  id: string;
  title: string;
  type: string;
  target: number;
  unit: string;
  deadline: string;
  progress: number;
  progressPercent: number;
  daysLeft: number;
}

export interface ActiveChallengeResult {
  challenge: ActiveChallenge | null;
}

export interface LeaderboardEntry {
  userId: string;
  name: string;
  username: string;
  /** @nullable */
  imageUrl: string | null;
  totalPoints: number;
  rank: number;
  isCurrentUser: boolean;
}

export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type WorkoutInviteStatus = typeof WorkoutInviteStatus[keyof typeof WorkoutInviteStatus];


export const WorkoutInviteStatus = {
  pending: 'pending',
  accepted: 'accepted',
  declined: 'declined',
  maybe: 'maybe',
} as const;

export interface WorkoutInviteRecord {
  id: string;
  senderId: string;
  receiverId: string;
  activity: string;
  location: string;
  date: string;
  time: string;
  status: WorkoutInviteStatus;
  createdAt: string;
}

export type WorkoutInvite = WorkoutInviteRecord & {
  senderName: string;
  receiverName: string;
};

export interface CreateWorkoutInviteRequest {
  receiverId: string;
  /** @minLength 1 */
  activity: string;
  /** @maxLength 200 */
  location?: string;
  /** @minLength 1 */
  date: string;
  /** @maxLength 50 */
  time?: string;
}

export type RespondToWorkoutInviteRequestStatus = typeof RespondToWorkoutInviteRequestStatus[keyof typeof RespondToWorkoutInviteRequestStatus];


export const RespondToWorkoutInviteRequestStatus = {
  accepted: 'accepted',
  declined: 'declined',
  maybe: 'maybe',
} as const;

export interface RespondToWorkoutInviteRequest {
  status: RespondToWorkoutInviteRequestStatus;
}

export interface MessageThread {
  friendId: string;
  friendName: string;
  /** @nullable */
  friendUsername: string | null;
  /** @nullable */
  friendImageUrl: string | null;
  lastMessage: string;
  /** @nullable */
  lastMessageAt: string | null;
  unreadCount: number;
}

export interface DirectMessageRecord {
  id: string;
  senderId: string;
  receiverId: string;
  content: string;
  read: boolean;
  createdAt: string;
}

export interface SendDirectMessageRequest {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type GetLeaderboardParams = {
limit?: number;
};