# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success (defaults to `/`) or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
  /dashboard/summary:
    get:
      tags: [Dashboard]
      operationId: getDashboardSummary
      summary: Get aggregated dashboard summary for the current user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Dashboard summary data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /dashboard/current-week:
    get:
      tags: [Dashboard]
      operationId: getCurrentWeekSchedule
      summary: Get this week's scheduled workouts (Sun–Sat) for the current user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Scheduled workouts for the current week.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeekScheduleEntry'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /dashboard/active-challenge:
    get:
      tags: [Dashboard]
      operationId: getActiveChallenge
      summary: Get the current user's most recent active challenge with progress
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Active challenge or null if none.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActiveChallengeResult'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /dashboard/leaderboard:
    get:
      tags: [Dashboard]
      operationId: getLeaderboard
      summary: Get the top users by total points for the leaderboard podium
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Ranked leaderboard entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'
        '401':
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    DashboardSummary:
      type: object
      required: [workoutsThisWeek, streakDays, caloriesToday, calorieGoal, proteinToday, carbsToday, fatToday, proteinGoal, carbGoal, fatGoal, totalPoints, name, imageUrl]
      properties:
        workoutsThisWeek:
          type: integer
        streakDays:
          type: integer
        caloriesToday:
          type: number
        calorieGoal:
          type: integer
        proteinToday:
          type: number
        carbsToday:
          type: number
        fatToday:
          type: number
        proteinGoal:
          type: integer
        carbGoal:
          type: integer
        fatGoal:
          type: integer
        totalPoints:
          type: integer
        name:
          type: string
        imageUrl:
          type: ["string", "null"]
    WeekScheduleEntry:
      type: object
      required: [id, date, title, completed]
      properties:
        id:
          type: string
        date:
          type: string
        title:
          type: string
        completed:
          type: boolean
        source:
          type: ["string", "null"]
    ActiveChallengeResult:
      type: object
      required: [challenge]
      properties:
        challenge:
          oneOf:
            - $ref: '#/components/schemas/ActiveChallenge'
            - type: 'null'
    ActiveChallenge:
      type: object
      required: [id, title, type, target, unit, deadline, progress, progressPercent, daysLeft]
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        target:
          type: number
        unit:
          type: string
        deadline:
          type: string
        progress:
          type: number
        progressPercent:
          type: number
        daysLeft:
          type: integer
    LeaderboardEntry:
      type: object
      required: [userId, name, username, imageUrl, totalPoints, rank, isCurrentUser]
      properties:
        userId:
          type: string
        name:
          type: string
        username:
          type: string
        imageUrl:
          type: ["string", "null"]
        totalPoints:
          type: integer
        rank:
          type: integer
        isCurrentUser:
          type: boolean
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...