# lib/api-zod/src/generated/types/activeChallenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActiveChallenge {
  id: string;
  title: string;
  type: string;
  target: number;
  unit: string;
  deadline: string;
  progress: number;
  progressPercent: number;
  daysLeft: number;
}