# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  workoutsThisWeek: number;
  streakDays: number;
  caloriesToday: number;
  calorieGoal: number;
  proteinToday: number;
  carbsToday: number;
  fatToday: number;
  proteinGoal: number;
  carbGoal: number;
  fatGoal: number;
  totalPoints: number;
  name: string;
  /** @nullable */
  imageUrl: string | null;
}