# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  userId: string;
  name: string;
  username: string;
  /** @nullable */
  imageUrl: string | null;
  totalPoints: number;
  rank: number;
  isCurrentUser: boolean;
}