# lib/api-zod/src/generated/types/weekScheduleEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeekScheduleEntry {
  id: string;
  date: string;
  title: string;
  completed: boolean;
  /** @nullable */
  source?: string | null;
}